refactor/rename: 각 step 컴포넌트 공통 안내 타이틀 컴포넌트로 분리, common 폴더 createPortfolio 폴더로 공통 컴포넌트 파일 이동
